+++ class/helios/HeliosStatusSQL.class.php
@@ -2,6 +2,7 @@
 
 class HeliosStatusSQL {
 
+	const ERREUR = -1;
 	const STATUS_EN_ATTENTE_TRANMISSION_SAE = 19;
 	const STATUS_ERREUR_LORS_DE_L_ENVOI_SAE = 20;