+++ test/PHPUnit/class/helios/HeliosEnvoiControlerTest.php
@@ -34,6 +34,9 @@ class HeliosEnvoiControlerTest extends S2lowTestCase {
 		parent::tearDown();
 	}
 
+	/**
+	 * @throws Exception
+	 */
 	public function testValidateAllTransactions(){
 		$this->validatePesAller("pes_aller_ok.xml");
 		$authoritySiret = new AuthoritySiretSQL($this->getSQLQuery());
@@ -41,6 +44,11 @@ class HeliosEnvoiControlerTest extends S2lowTestCase {
 		$this->assertEquals("12345678912345",$siret_list[0]['siret']);
 	}
 
+	/**
+	 * @param $filename
+	 * @return array|bool|mixed
+	 * @throws Exception
+	 */
 	private function validatePesAller($filename){
 		$pes_aller = __DIR__."/../../helios/fixtures/{$filename}";
 		copy($pes_aller,$this->testStreamUrl."/helios/".sha1_file($pes_aller));
@@ -51,7 +59,22 @@ class HeliosEnvoiControlerTest extends S2lowTestCase {
 		return $id_t;
 	}
 
+	/**
+	 * @throws Exception
+	 */
+	public function testNotInIso8859(){
+		$id_t = $this->validatePesAller("pes_aller_utf8.xml");
 
+		$heliosTransaction = new HeliosTransactionsSQL($this->getSQLQuery());
+		$info = $heliosTransaction->getInfo($id_t);
+		$this->assertEquals(HeliosStatusSQL::ERREUR,$info['last_status_id']);
+		$last_status_info = $heliosTransaction->getLastStatusInfo($id_t);
+		$this->assertEquals("Transaction $id_t : ce fichier n'est pas encodé en ISO-8859-1",$last_status_info['message']);
+	}
+
+	/**
+	 * @throws Exception
+	 */
 	public function testAccentNomFic(){
 		$id_t = $this->validatePesAller("PESALR2_accent_dans_nomfic.xml");
 
@@ -65,6 +88,9 @@ class HeliosEnvoiControlerTest extends S2lowTestCase {
 		$this->assertEquals("12345678912345",$siret_list[0]['siret']);
 	}
 
+	/**
+	 * @throws Exception
+	 */
 	public function testRetrieveAllPesInfo(){
 		$id_t = $this->validatePesAller("pes_aller_ok.xml");
 		$heliosTransaction = new HeliosTransactionsSQL($this->getSQLQuery());
@@ -75,10 +101,16 @@ class HeliosEnvoiControlerTest extends S2lowTestCase {
 		$this->assertEquals("034000",$info['xml_id_post']);
 	}
 
+	/**
+	 * @throws Exception
+	 */
 	public function testSendSamePESAller(){
 		$this->sendSamePESAllerFailed();
 	}
 
+	/**
+	 * @throws Exception
+	 */
 	private function sendSamePESAllerFailed(){
 		$this->validatePesAller("pes_aller_ok.xml");
 		$id_t = $this->validatePesAller("pes_aller_ok.xml");
@@ -90,6 +122,9 @@ class HeliosEnvoiControlerTest extends S2lowTestCase {
 	}
 
 
+	/**
+	 * @throws Exception
+	 */
 	public function testSendSamePESAllerDoNotVerify(){
 		$this->heliosEnvoiControler->setDoNotVerifyNomFicUnicity(true);
 		$authoritySQL = new AuthoritySQL($this->getSQLQuery());
@@ -101,11 +136,17 @@ class HeliosEnvoiControlerTest extends S2lowTestCase {
 		$this->assertEquals(2,$info['status_id']);
 	}
 
+	/**
+	 * @throws Exception
+	 */
 	public function testSendSamePESAllerDoNotVerifyOnlyConst(){
 		$this->heliosEnvoiControler->setDoNotVerifyNomFicUnicity(true);
 		$this->sendSamePESAllerFailed();
 	}
 
+	/**
+	 * @throws Exception
+	 */
 	public function testSendSamePESAllerDoNotVerifyOnlyAuthority(){
 		$authoritySQL = new AuthoritySQL($this->getSQLQuery());
 		$authoritySQL->updateDoNotVerifyNomFicUnicity(1,true);