+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Le script cron/jour.php est remplacé par le script script/actes/actes-send-classification-for-all-authorities.php #308
 - Si un message métier d'anomalie est reçu alors que la transaction n'est pas à l'état transmis, on ne change pas l'état de la transaction et on met le message dans la banette des erreurs #305
 - Possibilité de spécifier le type de la pièce principale sur le traitement par lot #318
+- Refus des PES Aller qui ne sont pas en ISO-8859-1 #373
 
 ### Corrections